pyproject.toml
[tool.poetry]
name = "platform-proforall"
version = "0.1.0"
description = ""
authors = ["Михаил <misga.yats@gmail.com>"]
readme = "README.md"
packages = [{include = "platform_proforall"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
pillow = "^10.0.0"
django-filter = "^23.2"
rest-framework-simplejwt = "^0.0.2"
coverage = "^7.2.7"
requests = "^2.31.0"
stripe = "^5.5.0"
drf-yasg = "^1.21.7"
celery = "^5.3.1"
redis = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

